OPAL-2748 Async R session operations are done in a thread associated with a detached shiro session
